requirements.txt
altgraph==0.17
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
future==0.18.2
idna==2.8
jdcal==1.4.1
mccabe==0.6.1
openpyxl==3.0.3
pefile==2019.4.18
pycodestyle==2.5.0
pyflakes==2.1.1
PyInstaller==3.6
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
yapf==0.29.0
